Pineapple Island is a real estate website
Pineapple Island is the most sought-after community in Ibadan that values and puts the needs of their clients first. It is a territory with loft complexes that creates a country life and homely feeling that is integrated with industrial development. A perfect, professional built abode that gives you the true taste of beauty and happiness.